requirements.txt
python-dotenv
adk
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-aiplatform
google-cloud-storage
google-adk
fastapi
uvicorn




